Don't let KindOfUninit pass nullable VerifyParamTypes
Nullable typehints that are nullable because of defaulted
parameters should still work the same with this, because the DV init
entry point for the parameter will set the local to an initialized
null before jumping back to the main entry point (which has the
VerifyParamType opcodes).  This changes runtime behavior for ?Foo
hints, but hopefully nothing is relying on being able to omit the
argument entirely right now.

Reviewed By: @elgenie

Differential Revision: D1131777