Fix Cns ir opcode output types
Summary: According to http://us3.php.net/define, constants can be null, bool,
int, float, string, and resource. The change itself was pretty simple but it
exposed a bug in the refcount opts where DefLabels weren't being properly
handled. After much thought and discussion, we decided the best way to handle
this is to require DefLabel to indicate which of its destinations should be
forwarding unconsumed references from their sources. See the comment on
IRBuilder::cond for more information.

Reviewed By: @jdelong

Differential Revision: D1294548