Update Class/Func/Unit layout based on data profiling
Based on the load address profiling thats possible with
the 3.10 kernel, I've reorganized the fields of Class, Func and
Unit to try to keep the hottest fields close together.

Perflab doesn't show much in terms of cpu time, but it seems to
help quite a lot (on the order of 1-2%) with dCache, L3-cache and
iTLB misses.

Reviewed By: @edwinsmith

Differential Revision: D1195077