Improve couldBeThisObj/mustBeThisObj for hhbbc minstr inference
A TODO was to look a the type of bases that aren't inside the
frame.  I did this, but then I realized neither of these changes can
kick in yet.  The couldBeThisObj needs support for unrelated object
types knowing they aren't related (#3343798), and the mustBeThisObj
will always say false if the base isn't in the FrameThis anyway,
because anything we know about a heap type must include InitNull at
least right now.

We might as well put this in as a test case for #3343798 though.  (And
I clarified some comments about how refs work, and fixed something
wrong with BaseR wrt this.)

Reviewed By: @dariorussi

Differential Revision: D1135320