Make use of hhbbc's static property types in the JIT
Add a RepoAuthType to each SProp, and use it in the JIT if
it's there.

This code was easiest to manipulate after removing the code sharing
between the FooS and FooG instructions.  I considered putting them
back together after that, but I think since it was easier to work with
without it and it wasn't sharing all that much, it's probably better
to leave it this way.  I'm cleaning up the FooG functions in a diff on
top of this for easier review.

Reviewed By: @bertmaher

Differential Revision: D1137138