Script to execute may be outside DOCUMENT_ROOT
Php files are often inside the docroot, but that is not
a requirement of the apache config.  Php files may be aliased
into the docroot.  Accept PATH_TRANSLATED when it is within
the docroot or within the Server.SourceRoot runtime param so
these environments can continue to work similar to how they
worked with php-fpm.