If we can constant propagate, make it imply it can't throw
Make use of this when deciding whether to propagate across a
factored edge to an exception handler.  Diff includes a script that
now infers the "impossible" block is unreachable, but didn't without
the change.  It's probably not much of a test case for this (it would
pass without this diff), but it is a decent edge case in exception
correctness so we might as well check it in.

Reviewed By: @bertmaher

Differential Revision: D1137528