Simplify region selector/pgo options, fix pgo for legacy/tracelet selectors
This diff separates the JitPGO and JitRegionSelector runtime options
and adds a JitPGORegionSelector option. JitPGO can now be toggled for the
legacy and tracelet region selectors, and the hottrace and hotblock region
selectors are the choices for JitPGORegionSelector. It also has the required
tweaks to the tracelet region selector to select appropriately-sized regions
for profiling (the legacy selector gets it for free since it starts with a
Tracelet from analyze).

Reviewed By: @ottoni

Differential Revision: D1139217