Tracelet region selector tweaks
This diff adds an option to compare the RegionDescs generated by the
legacy and tracelet selectors. I then used it with test/{quick,slow} to find
and fix a number of differences. Most were related to guard relaxation,
including fairly major changes to how it uses AssertType ops to eliminate
guards when possible.

Reviewed By: @jdelong

Differential Revision: D1151684