Don't return any RepoAuthTypes unless we're in WholeProgram mode
Bsimmers found a case where hphpc says something has
KindOfString (implying InitNull or String), but it could be
KindOfUninit.  In whole program mode, hphpc correctly finds out it
could be unset, but not in normal mode.  I didn't really try to dig in
to why, but let's just disable it unless we're in RepoAuthoritative
mode.

Reviewed By: @swtaarrs

Differential Revision: D1144005