Modify emitter to always produce 1:1 pairs of inner/outer generator functions
I want to make it an invariant for now that inner generators
always have a unique outer generator.  This is probably temporary,
since @jano has plans to combine these functions eventually.  At
runtime we clone the body for each class context it can run in
anyway---it makes sense to do it statically so we can annotate
statically-known types based on the context, too.

Also, depending on how hphpc handles type inference for closure
generators in classes, it might even be incorrect not to do this
earlier (on the AST), since the same bytecode body can run with
multiple class contexts (i.e. it is one function that might have
access to private properties in several different classes).  I'm not
sure if we're handling that right or not ...

Reviewed By: @paroski

Differential Revision: D1144199