Describe some basics of closures, generators, and async functions in spec
We don't really have anything about classes yet either, so I
put something very small about that for now.  The async functions
blurb is really light, because there are plans to stop using these
"generator function pairs" for them soon.  This at least gives us a
place to list things we are relying on about the implementation for
now, though.

Reviewed By: @dariorussi

Differential Revision: D1144201