Add an option to match namespaces when not specifying one.
This restores functionality present in HHVM's previously "broken"
SimpleXML implementation but apparently used throughout Facebook's code.

I wasn't sure on where to put the config option, so I created
a separate SimpleXML node just in case we needed more of these in the
future.

Don't forget to -vSimpleXML.EmptyNamespaceMatchesAll=true to enable
this.