Make ext_datetime/strtotime_leak.php less sensitive to JIT-time allocations
When we're jitting, we use the smart allocator, which affects
what memory_get_usage will say.  Changing tracelet shapes (or
inserting a bunch of extra assert type instructions, as hhbbc does)
can push this over the limit for the leak it's trying to detect.  Just
make it a little less sensitive.

Reviewed By: @edwinsmith

Differential Revision: D1147403