PATH_TRANSLATED is a full path to the script
Utilize the new m_scriptFilename to handle PATH_TRANSLATED for apache.
This removes the requirement that PATH_TRANSLATED must be prefixed
with DOCUMENT_ROOT, which is utilized by some installations.

To support the principle of least suprise this duplicates the condition
used by php-fpm to determine if PATH_TRANSLATED should be used as its
full path.