fix apache mod_fastcgi
It turns out the SCRIPT_FILENAME was the wrong thing to trust. That is the config option that is used for nginx, but it isn't what goes on the wire. In nginx, apache mod_fastcgi and mod_fastcgi_proxy all seem to send this as the absolute path, so lets use it.

Closes #1663

Reviewed By: @scannell

Differential Revision: D1148606