start of INI revoluation
We have a bajillion runtime options. We should cut them into 3 categories:

* Apache or Nginx setting replacements
* Exact mirrors of php ini settings
* New things that we think should be configurable

This diff tackles the second category. I went through a bunch of the manual pages for ini settings and tried to find the same setting in our hdf. While I was doing that I figured out that exposing some of those to userland would be a security problem, so I introduced the PHP idea of where a setting can be set. The only place I limited it was the `ini_set` call for now, but we can build out more if we support per-dif `php.ini`.

Then while I was doing this, I did lots of stupid things like hooking the wrong variable up to the wrong setter and getter, so I made the compiler make the choice for me by overloading `Bind` a bunch.

So yeah, we got 3 things rolled into this diff (new ini settings, ini setting security, and ini type inference) but I think they jive well together and cutting them apart will be a bit painful.

Reviewed By: @fredemmott

Differential Revision: D1149351