Rename SplObjectStorage::$storage to $__storage
Users can set arbitrary properties on these objects. This is perfectly
fine in Zend:

  <?php

  $o1 = new stdClass();
  $store = new SplObjectStorage;
  $store->attach($o1);
  $store->storage = 'herpderp';

We still have a problem if someone chooses to set $store->__storage, but as double-underscore is a fairly common 'this is private' indicator, that should be fine. Could alternatively rename to __hhvmStorage if that's a concern - don't see how to get rid of the underlying issue though.

Reviewed By: @ptarjan

Differential Revision: D1149297