Change Code Model to allow boolean properties to default to false if not specified.
The few instances of boolean properties in the Code Model are almost always false, so it makes it easier to serialize and construct a model if these can be omitted. Changed the class generators to treat null booleans as false. Also changed the serializer to omit false property values from the serialization. Fixed the serialization of object property and object method expressions to serialize property/method name as a string rather than a scalar values. Now also serialize the type annotation of parameter declarations.

Reviewed By: @shai-d

Differential Revision: D1115222