Some progress on getting rid of Type::None
This is largely mechanical. IRInstruction's type param is now a
folly::Optional<Type> internally, though it still exposes the same
interface (folly::Optional will assert if you try to get the value of an
Optional that holds no value). The network of functions for InterpOne in
hhbc-translator now use Optional too.

Reviewed By: @jdelong

Differential Revision: D1151593