Don't require an explicit framework class
Right now, we require an explicit class definition for all frameworks that will be run by the test runner. This is true even if the class is just a cookie cutter call to the Framework class default constructor.

Make this not be the case anymore. Only require an explicit class for a specific framework if default behavior needs to be overriden.

Reviewed By: @alexmalyshev

Differential Revision: D1151499