Get old idl-style "native" units into hhbbc's list of UnitEmitters
We merge the native units during ProcessInit, which is too
early for hhbbc.  It appears to be harmless to create the unit
emitters again, so restructure this to do that and pass it to hhbbc.
Non-hhbbc shouldn't be affected.  Eventually I want to change this
code so hhbbc has its own main() and operates on a repo, so I hope
folks can pardon it being a little gross for now.

Reviewed By: @dariorussi

Differential Revision: D1152464