Rewrite PHINQ select clauses into client parts and server parts.
LINQ select clauses can contain arbitrary source language expressions. These cannot always be evaluated on the query server, so the clauses passed to the query provider need to be sanitized and the query provider needs to be given a way to turn the tuples that come from the query server into the values specified by the original select clause. This is done by rewriting query expressions during parse time so that the serialized code model for the query expression is sanitized and a new lambda function and associated closure expression is constructed for use in the code generated by the query expression.

The overall scheme is now something like this: Given
  $r = FROM $r in $t WHERE $r->c == $someValue SELECT new SomeObject($r->d, ($r->e).'foo');
we end up generating roughly
  $r = $t->executeQuery("FROM $r in __query_param0__ WHERE $r->c == __query_param1__ SELECT tuple($r->d, $r->e);", $__query_result_row__ ==> new SomeObject($__query_result_row__->r::d, $__query_result_row__->r::e.'foo'), $someValue);

The value that ends up in $r is up to the query provider, but the expectation is that whenever and however it ends up producing result elements, they are turned into SomeObject instances via calls to the provided closure.

Reviewed By: @paroski

Differential Revision: D1151254