Use a NameA opcode for implementing ::class
There are no class constants named "class", so either ClsCns
needed a documentation update or it was not behaving as specified.
This should also "fix" the self::class in a pseudo-main to fail the
way parent::class does, matching zend.  static::class still works in
this case, although it doesn't in zend.

Reviewed By: @ptarjan

Differential Revision: D1154554