Support to infer return types and parameter reffiness for overridden methods
Adds a res::Func mode that represents a "FuncFamily", which
is just a set of possible functions it could call.  Return types and
parameter reffiness in this mode turns into loops over the possible
functions.

This code is disabled because it causes a perf regression.

To compute the possible functions in a family, when the index is
initially built, this adds a pass to instantiate all possible
combinations of the classes so we can see all possible subclasses.
This could have other uses later, too.  (This costs about 6 seconds
for create_index on www in dbg so I think it's ok not to parallelize
or whatnot for now.)

Reviewed By: @dariorussi

Differential Revision: D1157919