allow all runtime settings to be set by ini files
I couldn't find a single member of the team that liked HDF files (except maybe @markw65 who was pretty indifferent a few months back). For PHP compat we must support INI files anyways. Having 2 config formats for the same things is bad pracice. These 3 points seem to make migrating out of HDF into INI a no-brainer. Now how to do it?

I'm planning that every HDF option is flattened down to an ini option prefixed with `hhvm.` and converted to underscores instead of camel case.

The migration path is to support all options to be settable by both ini files and hdf files. Then move everything over to ini files, then delete the hdf support.

I changed the `-c` option to support either hdf or ini. Then I moved all the important logic from `hdf.cpp` to `RuntimeOptions`. `::Bind` is the most important for now. The final diff in this series will remove `hdf` as an argument to any of the functions and to all the parsing routines.

A few decisions I made along the way:
* The main API is `Bind(Variable,...)` instead of `Variable = GetBool()`. I like the compiler to decide what type you want. This makes it a bit awkward to replace `bool foo = hdf["thing"].getBool();` so there is a `GetBool` for temporaries, but please use it sparely as it won't actually make an ini setting.
* I used `folly:dynamic` as the container for INI settings and put the parsing logic on RuntimeOptions. I'm open to other ideas here if we want a dedicated parser or holder.
* I'm not using an INI section header, since PHP ignores those when parsing `php.ini` files, I want our files to be parsable by vanilla php too.
* I was pretty careful with defaults, always verifying that the statically set one exactly matched the one that was really used, which was passed into the HDF parser.

Reviewed By: @fredemmott

Differential Revision: D1155179