Only rename builtins that aren't C++ builtins to __builtin_name
It looks like the only case in systemlib that does this
fallback stuff is array_map.  (I believe array_filter used to.)  Since
this function was detecting that situation by trying to look up loaded
functions, it would rename every function if you call it a second time
to create the UnitEmitter again (which I need for HHBBC).  This way
fixes that, but leaves the dependency on the hhas system being loaded
before the native func one.

Reviewed By: @markw65

Differential Revision: D1155030