Fix bugs in query op predicates introduced by removal of JmpIs{N,}Type
This fusion was apparently actually happening after all, at
least with hhbbc turned on.  Maybe we'll want to implement something
like it at some point (or use a generic branch fusion thing), but
right now it is probably best to leave it off, because it wasn't
implemented in jump-opts and the REQ_BIND_JMPCC would end up in
astubs, which seems worse than one extra setcc and test opcode.

Reviewed By: @bertmaher

Differential Revision: D1155023