Don't allow unserialize to break property types in RepoAuthoritative mode
It's possible to crash the VM by unserializing strings that
give object properties types that we can't see statically.  Before
hhvm this was apparently handled by the smart pointer type rejecting
invalid unserialization, but now we just crash if it happens.  This is
still a work in progress, so it's behind a runtime option that's off
for now.  (We might need to pre-resolve persistent Class*'s from the
names in the RepoAuthType to make this performant.)

Reviewed By: @paroski

Differential Revision: D1156613