Fix a few hhbbc bugs with Uninit, and a horrible bug in IncDecL
Fix a bug relating to the possibility of accessing $this->foo
after foo was unset on a class with a magic getter.  Improve some
fooAsCell methods.  Also fix a horrible bug in IncDecL.  Still a TODO
here is being smarter in thisPropAsCell and checking if the current
class could actually have a __get before pessimizing, but it should be
correct wrt this at least now.

Reviewed By: @dariorussi

Differential Revision: D1163454