Fixes for Code Model for query expressions.
The serialization code for query expressions did not match the actual code model in a various places. Also, the model for into expressions was incorrect. Also changed the way that order by AST nodes are constructed because the tokens for ascending and descending always report 0 as their token id value.

Reviewed By: @paroski

Differential Revision: D1157921