Add control of number of threads and running like PHPUnit
Add two options:

1. To specify the number of threads the runner should use run running a set
of framework tests.

2. To allow one to specify that framework tests should be run just like they
would be with normal PHPUnit. This has the effect, essentially, of setting
the Framework $parallel member to false at the command line. This setting
could cause the testing to stop in its tracks with a fatal, just like
PHPUnit stops too.

Pear actually is --as-phpunit by default (with the override of $parallel = false)
in its constructor, because it breaks otherwise.

Reviewed By: @ptarjan

Differential Revision: D1165221