Make hhbbc parallelism options configurable
Move things into a namespace.  It may seem strange to
configure this using global variables---but we better "own" the
behavior of the whole process when using these kinds of functions or
something is already wrong, so it seems to make sense.

Different amounts of parallel-ness may make sense for different things
at some point, but currently it seems fine to use the same settings
for everything it does in parallel.

Reviewed By: @dariorussi

Differential Revision: D1165675