Fix -fno-constant-prop
The type system basically always tries to propagate
constants, and the ConstantProp flag controls whether to change
bytecode based on knowledge of what's constant.  If we don't do the
second step, strings and arrays we thought would be static won't be,
so it's wrong to propagate the types them as if they were.

It seemed a little strange to me that the analysis passes basically
need to know what changes will be made later to infer types correctly,
but after thinking about it for a while it doesn't seem that
unreasonable in this case.  (Other solutions to this would require
re-doing the analysis after making bytecode changes to actually
propagate constants, which could lead to more constants to propagate,
etc ...)

Reviewed By: @edwinsmith

Differential Revision: D1165713