Handle async functions slightly better for class-at-a-time analysis
The interpreter doesn't understand certain
continuation-related opcodes and kills local types.  This was causing
some property types to turn into InitCell unnecessarily.

We more easily can have the types in the eager function, and the way
we probably want to handle async functions (until they are merged) is
by doing all the inference on the eager version.  It also was doing
killLocals on AsyncESuspend, but in practice I don't think that one
matters (it always comes right before a RetC, so the JIT should
hopefully know it doesn't have to decref anything).

Reviewed By: @dariorussi

Differential Revision: D1165785